compiler: Support module alongside source_relative
The natural behavior should the user want to specify a module but also
leverage source_relative paths output is to consider the module to be
the prefix for all import paths written to the generated go code.

In a `module example.com/proto/gen`, it's possible to do:

    $ tree
    go.mod
    go.sum
    src
      + types
        + foo.proto
        + bar.proto
      + widget
        + service.proto
    gen

    $ protoc $(find src -name "*.proto")
        --proto_path=src --go_out=gen
        --go_opt=paths=source_relative,module=example.com/proto/gen

    $ tree gen
    gen
      + types
        + foo.pb.go
	+ bar.pb.go
      + widget
        + service.pb.go

without specifying the `"example.com/proto/gen/..."` module prefix in
the acutal proto files. In order to do this with the `M` option, you
would have to build the package list manually and pass a dedicated `M`
option for each source proto. This gets rather cumbersome.

How is this different than the deprecated import_path option? The use
case solved here is very specific to a common pattern of generating
Go sources from a shared set of proto files into a single root module
path while still retaining the source directory/package structure. This
use case is simple enough and useful enough that it makes sense to
support it specifically without reverting to trying to support a fully
generalized import_path type of option that likely caused more confusion
than desired. It also makes sense of a previously incompatible option
set which is an aesthetically welcome reduciton in error output.

Related discussion: https://github.com/golang/protobuf/issues/1158